test: split channel restart stress tests
